requirements.txt
streamlit
pydantic
python-dotenv
numpy
scikit-learn
rapidfuzz
httpx

